chore: extract controller
